added login and registration
