--- Registration_and_Login/PRG3781_Project_Group6/README.md
@@ -1,2 +1,0 @@
-# PRG3781_Project_Group6
-Student Portal using Java Spring Boot
