--- Registration_and_Login/PRG3781_Project_Group6/appuser/AppUserRole.java
@@ -1,6 +1,0 @@
-package com.example.demo.appuser;
-
-public enum AppUserRole {
-    USER,
-    ADMIN
-}
