--- Registration_and_Login/PRG3781_Project_Group6/email/EmailSender.java
@@ -1,5 +1,0 @@
-package com.example.demo.email;
-
-public interface EmailSender {
-    void send(String to, String email);
-}
