--- Registration_and_Login/PRG3781_Project_Group6/registration/EmailValidator.java
@@ -1,14 +1,0 @@
-package com.example.demo.registration;
-
-import org.springframework.stereotype.Service;
-
-import java.util.function.Predicate;
-
-@Service
-public class EmailValidator implements Predicate<String> {
-    @Override
-    public boolean test(String s) {
-//        TODO: Regex to validate email
-        return true;
-    }
-}
