--- Registration_and_Login/PRG3781_Project_Group6/registration/RegistrationController.java
@@ -1,23 +1,0 @@
-package com.example.demo.registration;
-
-import lombok.AllArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping(path = "api/v1/registration")
-@AllArgsConstructor
-public class RegistrationController {
-
-    private final RegistrationService registrationService;
-
-    @PostMapping
-    public String register(@RequestBody RegistrationRequest request) {
-        return registrationService.register(request);
-    }
-
-    @GetMapping(path = "confirm")
-    public String confirm(@RequestParam("token") String token) {
-        return registrationService.confirmToken(token);
-    }
-
-}
