--- Registration_and_Login/PRG3781_Project_Group6/registration/token/ConfirmationTokenService.java
@@ -1,27 +1,0 @@
-package com.example.demo.registration.token;
-
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
-import java.util.Optional;
-
-@Service
-@AllArgsConstructor
-public class ConfirmationTokenService {
-
-    private final ConfirmationTokenRepository confirmationTokenRepository;
-
-    public void saveConfirmationToken(ConfirmationToken token) {
-        confirmationTokenRepository.save(token);
-    }
-
-    public Optional<ConfirmationToken> getToken(String token) {
-        return confirmationTokenRepository.findByToken(token);
-    }
-
-    public int setConfirmedAt(String token) {
-        return confirmationTokenRepository.updateConfirmedAt(
-                token, LocalDateTime.now());
-    }
-}
