--- src/main/java/com/student/studentRegisterLogin/Student.java
@@ -1,63 +1,0 @@
-package com.student.studentRegisterLogin;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "student")
-public class Student {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    
-    private long student_id;
-    @Column(nullable = false, length = 25)
-    private String student_name;
-    @Column(nullable = false, unique = true, length = 45)
-    private String student_email;
-    private String student_password;
-    private String student_address;
-
-    public long getStudent_id() {
-        return student_id;
-    }
-
-    public void setStudent_id(long student_id) {
-        this.student_id = student_id;
-    }
-
-    public String getStudent_name() {
-        return student_name;
-    }
-
-    public void setStudent_name(String student_name) {
-        this.student_name = student_name;
-    }
-
-    public String getStudent_email() {
-        return student_email;
-    }
-
-    public void setStudent_email(String student_email) {
-        this.student_email = student_email;
-    }
-
-    public String getStudent_password() {
-        return student_password;
-    }
-
-    public void setStudent_password(String student_password) {
-        this.student_password = student_password;
-    }
-
-    public String getStudent_address() {
-        return student_address;
-    }
-
-    public void setStudent_address(String student_address) {
-        this.student_address = student_address;
-    }
-}
