+++ src/main/java/com/student/studentRegisterLogin/appuser/AppUserRepository.java
@@ -1,0 +1,24 @@
+package com.example.demo.appuser;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
+
+@Repository
+@Transactional(readOnly = true)
+public interface AppUserRepository
+        extends JpaRepository<AppUser, Long> {
+
+    Optional<AppUser> findByEmail(String email);
+
+    @Transactional
+    @Modifying
+    @Query("UPDATE AppUser a " +
+            "SET a.enabled = TRUE WHERE a.email = ?1")
+    int enableAppUser(String email);
+
+}
