+++ src/main/java/com/student/studentRegisterLogin/appuser/AppUserRole.java
@@ -1,0 +1,6 @@
+package com.example.demo.appuser;
+
+public enum AppUserRole {
+    USER,
+    ADMIN
+}
