+++ src/main/java/com/student/studentRegisterLogin/email/EmailSender.java
@@ -1,0 +1,5 @@
+package com.example.demo.email;
+
+public interface EmailSender {
+    void send(String to, String email);
+}
