+++ src/main/java/com/student/studentRegisterLogin/registration/EmailValidator.java
@@ -1,0 +1,14 @@
+package com.example.demo.registration;
+
+import org.springframework.stereotype.Service;
+
+import java.util.function.Predicate;
+
+@Service
+public class EmailValidator implements Predicate<String> {
+    @Override
+    public boolean test(String s) {
+//        TODO: Regex to validate email
+        return true;
+    }
+}
