+++ src/main/java/com/student/studentRegisterLogin/registration/RegistrationController.java
@@ -1,0 +1,23 @@
+package com.example.demo.registration;
+
+import lombok.AllArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping(path = "api/v1/registration")
+@AllArgsConstructor
+public class RegistrationController {
+
+    private final RegistrationService registrationService;
+
+    @PostMapping
+    public String register(@RequestBody RegistrationRequest request) {
+        return registrationService.register(request);
+    }
+
+    @GetMapping(path = "confirm")
+    public String confirm(@RequestParam("token") String token) {
+        return registrationService.confirmToken(token);
+    }
+
+}
