+++ src/main/java/com/student/studentRegisterLogin/registration/RegistrationRequest.java
@@ -1,0 +1,18 @@
+package com.example.demo.registration;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@AllArgsConstructor
+@EqualsAndHashCode
+@ToString
+public class RegistrationRequest {
+    private final String firstName;
+    private final String lastName;
+    private final String address;
+    private final String email;
+    private final String password;
+}
